backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI Policy Atlas Backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # FastAPI and ASGI server
    # Reference: https://fastapi.tiangolo.com/
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0", # ASGI server with standard extras for performance
    # Database
    # Reference: https://docs.sqlalchemy.org/en/20/orm/extensions/asyncio.html
    "sqlalchemy[asyncio]>=2.0.0", # SQLAlchemy 2.0 with async support
    "asyncpg>=0.30.0", # PostgreSQL async driver
    "alembic>=1.13.0", # Database migrations
    # GitHub API client
    "pygithub>=2.0.0",
    # Environment variables
    # Reference: https://pydantic-docs.helpmanual.io/usage/settings/
    "pydantic-settings>=2.6.0", # Settings management with Pydantic
    # CORS middleware
    "python-multipart>=0.0.12", # Required for form data parsing
    # Celery for background tasks and periodic scheduling
    # Reference: https://docs.celeryq.dev/
    "celery>=5.4.0", # Distributed task queue (includes beat scheduler)
    "redis>=5.0.0", # Redis broker for Celery
    "fastapi-debug-toolbar>=0.6.3",
]
